+++ src/main/java/nextstep/subway/line/application/FarePolicyHandler.java
@@ -2,7 +2,7 @@ package nextstep.subway.line.application;
 
 import nextstep.subway.line.domain.Fare;
 import nextstep.subway.line.domain.Money;
-import nextstep.subway.line.domain.policy.BaseFarePolicy;
+import nextstep.subway.line.domain.fare.policy.BaseFarePolicy;
 
 public interface FarePolicyHandler {
