+++ src/main/java/nextstep/subway/line/domain/fare/policy/AgeDiscountFarePolicy.java
@@ -1,9 +1,9 @@
-package nextstep.subway.line.domain.policy;
+package nextstep.subway.line.domain.fare.policy;
 
 import java.util.List;
 import nextstep.subway.line.domain.Fare;
 import nextstep.subway.line.domain.Money;
-import nextstep.subway.line.domain.policycondition.AgeDiscountFareCondition;
+import nextstep.subway.line.domain.fare.policycondition.AgeDiscountFareCondition;
 
 public abstract class AgeDiscountFarePolicy implements BaseFarePolicy {
