+++ src/main/java/nextstep/subway/line/domain/fare/policy/DistanceFarePolicy.java
@@ -1,14 +1,14 @@
-package nextstep.subway.line.domain.policy;
+package nextstep.subway.line.domain.fare.policy;
 
 import java.util.List;
 import nextstep.subway.line.domain.Fare;
 import nextstep.subway.line.domain.Money;
-import nextstep.subway.line.domain.policycondition.DistancePolicyCondition;
+import nextstep.subway.line.domain.fare.policycondition.DistancePolicyCondition;
 
 public abstract class DistanceFarePolicy implements BaseFarePolicy {
 
     @Override
-    public Money getCalculateFare(Fare fare, Money money) {//int distance, Money fare
+    public Money getCalculateFare(Fare fare, Money money) {
         for (DistancePolicyCondition each : getConditions()) {
             money = money.plus(each.calculateFare(fare.getDistance()));
         }
