+++ src/main/java/nextstep/subway/line/domain/fare/policycondition/DistancePolicyCondition.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain.policycondition;
+package nextstep.subway.line.domain.fare.policycondition;
 
 
 import nextstep.subway.line.domain.Money;
@@ -7,8 +7,8 @@ public interface DistancePolicyCondition {
 
     default Money calculate(int distance, int minDistance, int premiumDistance,
         int additionalFare) {
-        return Money.won((int) (
-            (Math.floor((double) (distance - minDistance - 1) / premiumDistance) + 1)
+        return Money.won(
+            (int) ((Math.floor((double) (distance - minDistance - 1) / premiumDistance) + 1)
                 * additionalFare));
     }
